#404 - Criando arquivo para actions do stackoverflow, criando arquivo d actions para importar funcionalidades para o bot
