+++ actions/tests/test_actions.py
@@ -1,0 +1,5 @@
+import pytest
+import os
+import sys
+import inspect
+import actions.actions
